Add voting page and track referrals

+++ webapp/titanembeds/blueprints/api/api.py
@@ -1,4 +1,4 @@
-from titanembeds.database import db, Guilds, UnauthenticatedUsers, UnauthenticatedBans, AuthenticatedUsers, GuildMembers, Messages, get_channel_messages, list_all_guild_members, get_guild_member, get_administrators_list, get_badges
+from titanembeds.database import db, Guilds, UnauthenticatedUsers, UnauthenticatedBans, AuthenticatedUsers, GuildMembers, Messages, get_channel_messages, list_all_guild_members, get_guild_member, get_administrators_list, get_badges, DiscordBotsOrgTransactions
 from titanembeds.decorators import valid_session_required, discord_users_only, abort_if_guild_disabled
 from titanembeds.utils import check_guild_existance, guild_accepts_visitors, guild_query_unauth_users_bool, get_client_ipaddr, discord_api, rate_limiter, channel_ratelimit_key, guild_ratelimit_key, user_unauthenticated, checkUserRevoke, checkUserBanned, update_user_status, check_user_in_guild, get_guild_channels, guild_webhooks_enabled, guild_unauthcaptcha_enabled, get_member_roles, get_online_embed_user_keys, redis_store
 from titanembeds.oauth import user_has_permission, generate_avatar_url, check_user_can_administrate_guild
@@ -6,6 +6,7 @@ from flask import Blueprint, abort, jsonify, session, request, url_for
 from flask import current_app as app
 from flask_socketio import emit
 from sqlalchemy import and_
+from urllib.parse import urlparse, parse_qsl, urlsplit
 import random
 import json
 import datetime
@@ -466,16 +467,27 @@ def user_info(guild_id, user_id):
             usr["badges"].append("discordbotsorgvoted")
     return jsonify(usr)
     
-@api.route("/webhook/discordbotsorg/vote/{}".format(config.get("discordbotsorg-webhook-secret", "")), methods=["POST"])
+@api.route("/webhook/discordbotsorg/vote", methods=["POST"])
 def webhook_discordbotsorg_vote():
     incoming = request.get_json()
     client_id = incoming.get('bot')
     if config["client-id"] != client_id:
         abort(401)
+    if request.headers.get("Authorization", "") != config.get("discordbotsorg-webhook-secret", ""):
+        abort(403)
     user_id = incoming.get("user")
     vote_type = incoming.get("type")
+    params = dict(parse_qsl(urlsplit(url).query))
     if vote_type == "upvote":
         redis_store.set("DiscordBotsOrgVoted/" + user_id, "voted", 86400)
     else:
         redis_store.delete("DiscordBotsOrgVoted/" + user_id)
+    referrer = None
+    if "referrer" in params:
+        try:
+            referrer = int(float(params["referrer"]))
+        except:
+            pass
+    DBLTrans = DiscordBotsOrgTransactions(int(float(user_id)), vote_type, referrer)
+    db.session.add(DBLTrans)
     return ('', 204)
